requirements.txt
azure-functions
azure-storage-blob
azure-cosmos
azure-ai-documentintelligence
requests
PyYAML
pdfminer.six
python-dateutil
jsonschema